@@ -4,13 +4,13 @@ description: A sample Flutter application demonstrating use of the what3words Da
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  what3words: ^3.0.3
+  what3words: ^3.1.0-nullsafety.1
 
 dev_dependencies:
   flutter_test:
