@@ -1,20 +1,19 @@
 name: what3words
 description: >-
   A Dart library to use the what3words v3 API. This library can either be used in native Dart or Flutter applications
-version: 3.0.4
+version: 3.1.0-nullsafety.1
 homepage: https://github.com/what3words/w3w-dart-wrapper
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
- chopper: ^3.0.2
- json_serializable: ^3.3.0
- json_annotation: ^3.1.0
- os_detect: ^1.0.0
+ chopper: ^4.0.0-nullsafety.0
+ json_serializable: ^4.1.0
+ json_annotation: ^4.0.1
+ os_detect: ^2.0.0-nullsafety.0
 
 dev_dependencies:
-  build_runner: ^1.9.0
-  chopper_generator: ^3.0.4
-  pedantic: ^1.9.0
-  test: ^1.14.4
+  chopper_generator: ^4.0.0-nullsafety.0
+  pedantic: ^1.11.0
+  test: ^1.16.8
