@@ -1,4 +1,5 @@
-import 'package:what3words/src/request/autosuggest_with_coordinates_request.dart';
+import 'package:what3words/src/response/api_response.dart';
+import 'package:what3words/src/response/response.dart';
 
 import 'autosuggest_options.dart';
 import 'abstract_builder.dart';
@@ -8,33 +9,33 @@ import 'request.dart';
 
 class AutosuggestRequest extends Request<Autosuggest> {
   final String input;
-  final String nResults;
-  final String focus;
-  final String nFocusResults;
-  final String clipToCountry;
-  final String clipToBoundingBox;
-  final String clipToCircle;
-  final String clipToPolygon;
-  final String inputType;
-  final String language;
-  final String preferLand;
+  final String? nResults;
+  final String? focus;
+  final String? nFocusResults;
+  final String? clipToCountry;
+  final String? clipToBoundingBox;
+  final String? clipToCircle;
+  final String? clipToPolygon;
+  final String? inputType;
+  final String? language;
+  final String? preferLand;
 
   AutosuggestRequest._builder(AutosuggestRequestBuilder builder)
       : input = builder._input,
-        nResults = builder._options.nResults,
-        focus = builder._options.focus,
-        nFocusResults = builder._options.nFocusResults,
-        clipToCountry = builder._options.clipToCountry,
-        clipToBoundingBox = builder._options.clipToBoundingBox,
-        clipToCircle = builder._options.clipToCircle,
-        clipToPolygon = builder._options.clipToPolygon,
-        inputType = builder._options.inputType,
-        language = builder._options.language,
-        preferLand = builder._options.preferLand,
+        nResults = builder._options?.nResults,
+        focus = builder._options?.focus,
+        nFocusResults = builder._options?.nFocusResults,
+        clipToCountry = builder._options?.clipToCountry,
+        clipToBoundingBox = builder._options?.clipToBoundingBox,
+        clipToCircle = builder._options?.clipToCircle,
+        clipToPolygon = builder._options?.clipToPolygon,
+        inputType = builder._options?.inputType,
+        language = builder._options?.language,
+        preferLand = builder._options?.preferLand,
         super(builder.api);
 
-  Future<Autosuggest> execute() async {
-    return await super.call(api.what3words().autosuggest, Autosuggest(), [
+  Future<Response<Autosuggest>> execute() async {
+    return await super.call(api.what3words().autosuggest, [
       input,
       nResults,
       focus,
@@ -51,9 +52,9 @@ class AutosuggestRequest extends Request<Autosuggest> {
 }
 
 /// Builder for `autosuggest` API requests
-class AutosuggestRequestBuilder extends AbstractBuilder<Future<Autosuggest>> {
+class AutosuggestRequestBuilder extends AbstractBuilder<Future<Response<Autosuggest>>> {
   final String _input;
-  final AutosuggestOptions _options;
+  final AutosuggestOptions? _options;
 
   AutosuggestRequestBuilder(What3WordsV3 api, this._input, this._options) : super(api);
 
@@ -61,7 +62,7 @@ class AutosuggestRequestBuilder extends AbstractBuilder<Future<Autosuggest>> {
   ///
   ///return an [Future<Autosuggest>] representing the response from the what3words API
   @override
-  Future<Autosuggest> execute() {
+  Future<Response<Autosuggest>> execute() {
     return AutosuggestRequest._builder(this).execute();
   }
 }
