@@ -1,4 +1,5 @@
 import 'package:what3words/src/request/request.dart';
+import 'package:what3words/src/response/response.dart';
 
 import '../../what3words.dart';
 import 'autosuggest_options.dart';
@@ -6,33 +7,33 @@ import 'abstract_builder.dart';
 
 class AutosuggestWithCoordinatesRequest extends Request<AutosuggestWithCoordinates> {
   final String input;
-  final String nResults;
-  final String focus;
-  final String nFocusResults;
-  final String clipToCountry;
-  final String clipToBoundingBox;
-  final String clipToCircle;
-  final String clipToPolygon;
-  final String inputType;
-  final String language;
-  final String preferLand;
+  final String? nResults;
+  final String? focus;
+  final String? nFocusResults;
+  final String? clipToCountry;
+  final String? clipToBoundingBox;
+  final String? clipToCircle;
+  final String? clipToPolygon;
+  final String? inputType;
+  final String? language;
+  final String? preferLand;
 
   AutosuggestWithCoordinatesRequest._builder(AutosuggestWithCoordinatesRequestBuilder builder)
       : input = builder._input,
-        nResults = builder._options.nResults,
-        focus = builder._options.focus,
-        nFocusResults = builder._options.nFocusResults,
-        clipToCountry = builder._options.clipToCountry,
-        clipToBoundingBox = builder._options.clipToBoundingBox,
-        clipToCircle = builder._options.clipToCircle,
-        clipToPolygon = builder._options.clipToPolygon,
-        inputType = builder._options.inputType,
-        language = builder._options.language,
-        preferLand = builder._options.preferLand,
+        nResults = builder._options?.nResults,
+        focus = builder._options?.focus,
+        nFocusResults = builder._options?.nFocusResults,
+        clipToCountry = builder._options?.clipToCountry,
+        clipToBoundingBox = builder._options?.clipToBoundingBox,
+        clipToCircle = builder._options?.clipToCircle,
+        clipToPolygon = builder._options?.clipToPolygon,
+        inputType = builder._options?.inputType,
+        language = builder._options?.language,
+        preferLand = builder._options?.preferLand,
         super(builder.api);
 
-  Future<AutosuggestWithCoordinates> execute() async {
-    return await super.call(api.what3words().autosuggestWithCoordinates, AutosuggestWithCoordinates(), [
+  Future<Response<AutosuggestWithCoordinates>> execute() async {
+    return await super.call(api.what3words().autosuggestWithCoordinates, [
       input,
       nResults,
       focus,
@@ -49,9 +50,9 @@ class AutosuggestWithCoordinatesRequest extends Request<AutosuggestWithCoordinat
 }
 
 /// Builder for `autosuggest-with-coordinates` API requests
-class AutosuggestWithCoordinatesRequestBuilder extends AbstractBuilder<Future<AutosuggestWithCoordinates>> {
+class AutosuggestWithCoordinatesRequestBuilder extends AbstractBuilder<Future<Response<AutosuggestWithCoordinates>>> {
   final String _input;
-  final AutosuggestOptions _options;
+  final AutosuggestOptions? _options;
 
   AutosuggestWithCoordinatesRequestBuilder(What3WordsV3 api, this._input, this._options) : super(api);
 
@@ -59,7 +60,7 @@ class AutosuggestWithCoordinatesRequestBuilder extends AbstractBuilder<Future<Au
   ///
   ///return an [Future<Autosuggest>] representing the response from the what3words API
   @override
-  Future<AutosuggestWithCoordinates> execute() {
+  Future<Response<AutosuggestWithCoordinates>> execute() {
     return AutosuggestWithCoordinatesRequest._builder(this).execute();
   }
 }
