@@ -1,7 +1,7 @@
 /// Represents an API error encapsulating the error code and message
 class APIError {
-  String code;
-  String message;
+  String? code;
+  String? message;
 
   APIError(this.code, this.message);
 }
