@@ -10,7 +10,7 @@ class HeaderInterceptor implements RequestInterceptor {
   static final String w3w_wrapper = 'X-W3W-Wrapper';
 
   final String _apiKey;
-  final Map<String, String> _headers;
+  final Map<String, String>? _headers;
 
   final String _userAgent;
 
@@ -30,9 +30,7 @@ class HeaderInterceptor implements RequestInterceptor {
       w3w_wrapper: _userAgent
     };
 
-    if (_headers != null) {
-      _headers.forEach((k, v) => headers[k] = v);
-    }
+    _headers?.forEach((k, v) => headers[k] = v);
 
     var newRequest = request.copyWith(headers: headers);
     return newRequest;
