@@ -13,7 +13,7 @@ import 'what3words_service.dart';
 ///Instances of the What3WordsV3 class provide access to Version 3 of the what3words API.
 class What3WordsV3 {
   final _defaultEndpoint = 'https://api.what3words.com/v3';
-  What3WordsV3Service _service;
+  late What3WordsV3Service _service;
 
   ///Get a new API manager instance.
   ///
@@ -40,7 +40,7 @@ class What3WordsV3 {
   }
 
   void _setupHttpClient(
-      String apiKey, String endpoint, Map<String, String> headers) {
+      String apiKey, String endpoint, Map<String, String>? headers) {
     _service = What3WordsV3Service.create(apiKey, endpoint, headers);
   }
 
@@ -96,7 +96,7 @@ class What3WordsV3 {
   ///[options] The autosuggest options and clippings, check available options here https://docs.what3words.com/api/v3/#autosuggest
   ///
   ///returns a [AutosuggestRequestBuilder] instance suitable for invoking a `autosuggest` API request
-  AutosuggestRequestBuilder autosuggest(String input, {AutosuggestOptions options}) {
+  AutosuggestRequestBuilder autosuggest(String input, {AutosuggestOptions? options}) {
     return AutosuggestRequestBuilder(this, input, options);
   }
 
@@ -109,7 +109,7 @@ class What3WordsV3 {
   ///[options] The autosuggest options, check available options here https://docs.what3words.com/api/v3/#autosuggest
   ///
   ///returns a [AutosuggestWithCoordinatesRequestBuilder] instance suitable for invoking a `autosuggest` API request
-  AutosuggestWithCoordinatesRequestBuilder autosuggestWithCoordinates(String input, {AutosuggestOptions options}) {
+  AutosuggestWithCoordinatesRequestBuilder autosuggestWithCoordinates(String input, {AutosuggestOptions? options}) {
     return AutosuggestWithCoordinatesRequestBuilder(this, input, options);
   }
 
@@ -122,7 +122,7 @@ class What3WordsV3 {
   ///[rank] The rank of the selected suggestion.
   ///[options] The autosuggest options used on the autosuggest or autosuggest-with-coordinates call.
   ///returns a [AutosuggestWithCoordinatesRequestBuilder] instance suitable for invoking a `autosuggest` API request
-  AutosuggestSelectionRequestBuilder autosuggestSelection(String rawInput, String sourceApi, String words, int rank, {AutosuggestOptions options}) {
+  AutosuggestSelectionRequestBuilder autosuggestSelection(String rawInput, String sourceApi, String words, int rank, {AutosuggestOptions? options}) {
     return AutosuggestSelectionRequestBuilder(this, rawInput, sourceApi, words, rank, options);
   }
 
