@@ -1,14 +1,13 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import 'response.dart';
 part 'language.g.dart';
 
 /// Encapsulates the list of supported what3words languages
 @JsonSerializable(explicitToJson: true)
-class Language extends Response<Language> {
+class Language {
   List<Languages> languages;
 
-  Language({this.languages});
+  Language({required this.languages});
 
   factory Language.fromJson(Map<String, dynamic> json) =>
       _$LanguageFromJson(json);
@@ -23,7 +22,7 @@ class Languages {
   String code;
   String name;
 
-  Languages({this.nativeName, this.code, this.name});
+  Languages({required this.nativeName, required this.code, required this.name});
 
   factory Languages.fromJson(Map<String, dynamic> json) =>
       _$LanguagesFromJson(json);
