@@ -3,21 +3,21 @@ import 'package:test/test.dart';
 import 'dart:io' show Platform;
 
 void main() {
-  var api = What3WordsV3(Platform.environment['W3W_API_KEY']);
+  var api = What3WordsV3(Platform.environment['W3W_API_KEY']!);
 
   test('twoInvalidCoordsTest', () async {
     var location = await api.convertTo3wa(Coordinates(-200, -200)).execute();
     expect(location.isSuccessful(), false);
 
-    var error = location.getError();
+    var error = location.error();
     expect(error, What3WordsError.BAD_COORDINATES);
   });
 
   test('validCoordsTest', () async {
-    var location =
+    var locationRequest =
         await api.convertTo3wa(Coordinates(51.520847, -0.19552100)).execute();
-    expect(location.isSuccessful(), true);
-
+    expect(locationRequest.isSuccessful(), true);
+    var location = locationRequest.data()!;
     expect('filled.count.soap', location.words);
     expect('GB', location.country);
 
